+++ packages/components-react/projects/nextjs/tests/vrt/specs/overview.vrt.ts
@@ -12,15 +12,3 @@ test.describe('overview', async () => {
     await expect(page.locator('#app')).toHaveScreenshot(`overview-${viewportWidthXXL}.png`);
   });
 });
-
-test.describe('overview notifications', async () => {
-  test(`should have no visual regression`, async ({ page }) => {
-    await page.goto('/overview-notifications');
-    await page.setViewportSize({
-      width: viewportWidthXXL,
-      height: await page.evaluate(() => document.body.clientHeight),
-    });
-    await page.mouse.click(0, 0); // click top left corner of the page to remove autofocus
-    await expect(page.locator('#app')).toHaveScreenshot(`overview-notifications-${viewportWidthXXL}.png`);
-  });
-});