+++ packages/components-vue/src/pages/ModalBasicPage.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+  import { PModal } from '@porsche-design-system/components-vue';
+</script>
+
+<template>
+  <PModal heading="Heading" open="true">Some Content</PModal>
+</template>