+++ packages/components-vue/src/pages/ModalPrefixedPage.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+  import { PModal, PorscheDesignSystemProvider } from '@porsche-design-system/components-vue';
+</script>
+
+<template>
+  <PorscheDesignSystemProvider prefix="my-prefix">
+    <PModal heading="Heading" open="true">Some Content</PModal>
+  </PorscheDesignSystemProvider>
+</template>