+++ packages/components/src/styles/colors.ts
@@ -10,6 +10,7 @@ type ThemedColor =
   | 'backgroundSurfaceColor'
   | 'backgroundSurfaceColorDarken'
   | 'backgroundSurfaceColorLighten'
+  | 'backgroundShadingColor'
   | 'contrastLowColor'
   | 'contrastMediumColor'
   | 'contrastHighColor'
@@ -51,6 +52,7 @@ const themeLight = {
   backgroundSurfaceColor: '#EEEFF2',
   backgroundSurfaceColorDarken: '#CBCED7',
   backgroundSurfaceColorLighten: '#FFFFFF',
+  backgroundShadingColor: 'rgba(1, 2, 5, 0.67)',
   contrastLowColor: '#D8D8DB',
   contrastMediumColor: '#6B6D70',
   contrastHighColor: '#535457',
@@ -90,6 +92,7 @@ const themeDark = {
   backgroundSurfaceColor: '#212225',
   backgroundSurfaceColorDarken: '#040405',
   backgroundSurfaceColorLighten: '#3E4045',
+  backgroundShadingColor: 'rgba(38, 38, 41, 0.67)',
   contrastLowColor: '#404044',
   contrastMediumColor: '#88898C',
   contrastHighColor: '#AFB0B3',